Disable IntelliJ IDEA detection of OSGi frameworks
The IntelliJ IDEA project configuarion is derived from Gradle, not
Maven.  But the OSGi material present is currently used only by Maven,
not Gradle.  If we let IntelliJ IDEA detect that we are using OSGi, it
reports dozens of warnings and errors.  Better to avoid OSGi + IntelliJ
entirely for now.  We can revisit this in the future if our Gradle build
becomes more OSGi-savvy.