Tighten up code for a bulk filtering operation
Previously we were collecting all nodes $A, then collecting nodes to
exclude $B, then removing $B from $A.  Better to simply filter on the
fly, as we iterate over all of the nodes in the first place.  Also
better to treat the filter as a proper set that can perform
containment checks instead of repeatedly iterating over its contents
item by item.