Don't copy an entire array just to iterate over its elements
Instead, use ArrayList::asList to create a List backed by the existing
array.  Strictly speaking, these are not equivalent:  the array-backed
list is read-only, whereas the list copied from an array is
modifiable.  But I strongly suspect that nobody is modifying the list
used here anyway, so we may as well avoid the O(n) copy operation.