Avoid calling Collection.toArray with pre-sized array
Opportunities found using an IntelliJ IDEA Java Performance
inspection.  Fixes applied using the same.  Although this commit
changes lots of code, I personally did very little typing to enact the
change.  It's unlikely that I introduced any bugs due to typos.  If we
trust the inspection itself, then the risk of new bugs due to this
change is small.

But is this change even a good idea?  Surprisingly, yes, it generally
is.  According to the description found in IntelliJ IDEA's inspection:

> There are two styles to convert a collection to an array: either
> using a pre-sized array (like c.toArray(new String[c.size()])) or
> using an empty array (like c.toArray(new String[0]).
>
> In older Java versions using pre-sized array was recommended, as the
> reflection call which is necessary to create an array of proper size
> was quite slow. However since late updates of OpenJDK 6 this call
> was intrinsified, making the performance of the empty array version
> the same and sometimes even better, compared to the pre-sized
> version. Also passing pre-sized array is dangerous for a concurrent
> or synchronized collection as a data race is possible between the
> size and toArray call which may result in extra nulls at the end of
> the array, if the collection was concurrently shrunk during the
> operation.
>
> This inspection allows to follow the uniform style: either using an
> empty array (which is recommended in modern Java) or using a
> pre-sized array (which might be faster in older Java versions or
> non-HotSpot based JVMs).

See also <https://stackoverflow.com/a/29444594/91929> for some
benchmarks.  Or if you have a free afternoon, read
<https://shipilev.net/blog/2016/arrays-wisdom-ancients/> for more
in-depth experimentation and discussion than you ever thought was
possible on this topic.  The high-order bit from that page is: "Bottom
line: toArray(new T[0]) seems faster, safer, and contractually
cleaner, and therefore should be the default choice now."