Allow yet another JarFile leak with only a warning
I'm not sure why Eclipse let this through previously, but right now
Eclipse is showing an error in SourceDirCallGraph.main, which is
exactly what we configured it to do.  Per wala/WALA#236, we currently
have no good solution for these JarFile leaks.  So treat this as only
a warning, just as we're doing elsewhere.