Use Gradle task configuration avoidance
This change in Gradle's task API has been incubating for a while, but
has been officially stable since Gradle 5.1.  The general idea is to
avoid configuring tasks until they are actually needed, and thereby to
improve build performance.  In practice, I'm not seeing any noticeable
difference.  Regardless, this is supposed to become the way all tasks
are created in the future, so we may as well get on board.