Improved keySet(), plus working entrySet() and values() methods
The implementation approach here is quite different from the keySet()
approach used previously.  The previous approach built a full-blown
HashSet, added all keys to that, and returned it.  This gives a fully
functional Set as a result of the keySet() call, but it's also rather
heavyweight.

The new approach builds upon AbstractSet to return a minimally
functional Set object.  This Set object is Iterable, and its Iterator
works by simply scanning over the array slots, one by one.  So
effectively the returned Set merely holds a reference to the original
SmallMap, and Iterators over this Set merely hold a reference to the
SmallMap plus an int recording the next slot.  This is very
lightweight.  It should perform quite well too for callers that merely
use keySet() to iterate over keys.  Performing more complex operations
on the key set is possible, but will be relatively inefficient since
it's really still just a linear sequence of keys, not a full hash
table.

The new entrySet() and values() methods use the same strategy of a
minimally functional Set that provides slot-scanning Iterators.  The
only difference among these three is what they fetch when asked for
the i'th "item":  the i'th key for keySet()'s Iterators, the i'th
value for values()'s Iterators, or the i'th (key, value) pair for
entrySet()'s Iterators.  That one point of difference among the three
is neatly factored out as a single method to be implemented in each of
three concrete subclasses of an abstract SlotIteratingSet<K> base
class.