Fix Android scope to work with multi-dex apks (#378)
* Fix Android scope to work with multi-dex apks

This fixes AndroidAnalysisScope.setAndroidAnalysisScope to
always use all dex files in an apk instead of using the
first one only. Further, the api level was set to Android Pie (28).

* Fixed setupAndroidAnalysisScope

The method was broken for single dex files. The DexFileModule
constructor with three arguments expected a container file
(.oat|.apk) instead of a .dex file (contrary to the javadoc).

This update uses the old implementation for .dex files and the new one
for container files.

* Finished Android multidex scope generation

- API level is optional. If omitted, it is automatically inferred.
- Update to dexlib-2.2.6