Don't assume that Comparable.compareTo returns -1
The only requirement is that it return a negative number when the
first item is smaller than the second.  That negative number need not
be -1.

Right now, this specific compareTo method is implemented such that it
will always use -1.  But that's not something we should rely on, and
there's certainly no benefit to doing so.

Found and fixed by IntelliJ IDEA.