IntelliJ IDEA inspection profile to discourage back-sliding
The inspections turned on here are ones I've recently made a directed
effort to resolve completely.  Thus, running code inspection with this
profile should find nothing to complain about.  Rerunning this profile
regularly on the "WALA-Controlled" file scope should discourage us
from reintroducing messy code of the kinds we've previously cleaned
up.

This is certainly not the *complete* set of inspections that we fully
comply with.  It's only the subset I've personally focused on
recently.  Other inspections could probably be turned on too, but I
haven't tried to systematically identify which ones.