IntelliJ IDEA inspection scope for just WALA-controlled files
This defines a subset of WALA source files that are actually under the
WALA project's control.  It excludes downloaded sources that we are
not in a position to modify.

The exclusion patterns here are not exhaustive, but they are good
enough for our current purposes.  Most importantly, they exclude
"JLex/Main.java", which contains some questionable Javadoc comments
that we would fix if they were under our control.  By excluding this
file from inspections, we can keep the Javadoc inspections fairly
aggressive for WALA's real source code without triggering JLex
warnings we can't fix.