Convert native builds to use Gradle's new C++ plugins
According to a Gradle blog post from January
2018 (https://blog.gradle.org/introducing-the-new-cpp-plugins>, these
plugins are the future of Gradle native build support.  They are now
recommended for general use as of Gradle
5.2 (https://docs.gradle.org/5.2/release-notes.html#recommend-c++-projects-use-new-c++-plugins).

An important implication of this change is that each of our three
native artifacts is now built in its own Gradle subsubproject.
Previously these artifacts were built as additional targets in
subprojects that also built Java artifacts.  However, per a thread in
the Gradle
Forums (https://discuss.gradle.org/t/subproject-with-both-java-and-c/30509),
and a bug report on the Gradle native
project (https://github.com/gradle/gradle-native/issues/352), the new
C++ plugins no longer allow mixing of Java and C++ in one subproject.
The preferred approach is for each subproject to have just one
artifact, so that's what we now do.