Require that kept nodes be subset of unpruned graph
In PrunedCallGraph, require that `keep` (the set of nodes to be kept)
be a subset of the nodes in the unpruned graph.  Check for this at
construction time, and raise an exception if violated.

The caller could potentially change either the `keep` set or the
unpruned graph after constructing the PrunedCallGraph instance.  If
they do that, then PrunedCallGraph might break, and the caller gets to
keep all the pieces.  We could be rechecking the "subset" relationship
in each and every place it is assumed to hold, but that seems
excessive.  So instead we just add a big warning in the constructor's
Javadoc documentation.

See <https://github.com/wala/WALA/pull/427#discussion_r259913855> for
the concluding piece of the discussion that led us to enact this
"subset" requirement.