Avoid DNS lookups when comparing URLs for equality
java.net.URL.equals() resolves hostnames using DNS so that it can
recognize equivalent URLs with different names for the same host.  But
this is overkill for our purposes.  Instead, compare URLs by comparing
their string representations.  Not as precise, but good enough for us,
and avoids inserting slow network operations into the middle of equality
checks.

Also avoid sets of URLs for the same reason.  Instead, keep sets of
the string representations of URLs.