Be more precise about tasks' inputs and outputs
A Copy or Sync task always treats its entire destination directory as
the task's output.  When we're copying or extracting a single file, it
is often more useful to treat just that one file as the output.  This
requires turning the Copy or Sync task into a generic task that calls
the Project.copy method.

Also for these tasks, if task A's inputs are given as some other task
B, that implicitly does two useful things: (1) it treats B's *outputs*
as A's inputs, and (2) it establishes a dependency of A upon B.  We
can't use this everywhere, but where we can, it's a nice
simplification.