Fix likely bug in array size calculation
Using "Math.min(0, other)" as the length of an array was surely not
intended here.  If "other" is zero or positive, then this would
allocate a 0-element array.  If "other" is negative, then this would
fail after trying to allocate a negative-sized array.

Evidently we have no tests that trigger this code.  I don't feel
well-informed enough to add a suitable test myself.  I'm guessing what
the correct size calculation should be based on context, but this is
presumably still untested.