Enable all IntelliJ IDEA inspections that find nothing
Some of these inspections check for problems we would truly want to
avoid.  For these, it is good to know that as of now, none of those
problems are present.  If future changes create instances of these
problems, IntelliJ IDEA will find them.

Some of these inspections may check for "problems" that we disagree
with or just don't care about.  No such instances appear in WALA now,
but if any arise in the future, then we can at least be aware of them
and make a well-considered decision to either fix the problems or
disable the inspections that we disagree with.

As of this commit, every IntelliJ IDEA inspection that is enabled
finds nothing, and every IntelliJ IDEA inspection that is disabled
finds nothing.

(Disclaimer #1: "every" as of IntelliJ IDEA 2019.1.  Future IntelliJ
IDEA releases will surely include new inspections.)

(Disclaimer #2: "every" inspection from standard IntelliJ IDEA or from
the plugins that I happen to have installed.  Additional plugins could
certainly define new inspections I haven't seen, and for those, they
will either be enabled or disabled depending on the plugin author's
decision for what the default behavior should be.)