Create a common "linters" task for extra checkers
These extra checkers are things we want to test for, but which are not
Java JUnit tests.  @msridhar has expressed a preference in the past
for not simply adding these extra checks as dependencies of the
"check" task.  So instead we now have a distinct "linters" task as a
single collection point for these other checks.