Port "change-version" script from Python to Bash
I was feeling an urge to validate our Python code somehow, such as via
flake8 or pychecker.  However, we only had one Python script in the
entire project, and mostly what it did was run other subcommands.

The easier path to validation, then, was to port this script from
Python to Bash so that it can be checked by our existing shellCheck
task.  The Bash version ends up being a few lines shorter, too.