Convert to Gradle's java-library plugin
The "java-library" plugin distinguishes dependencies needed only for a
library's internal implementation from dependencies that form part of
a library's external API.  A library's implementation dependencies are
not transitively made available to other components that depend on
that library.  This keeps classpaths shorted and helps reduce
recompilation.  See
<https://docs.gradle.org/current/userguide/java_library_plugin.html>
for more information.

I took an empirical approach to determining which dependencies are
implementation dependencies and which are API dependencies.  In
general, one should prefer the more restrictive implementation
categorization wherever possible.  What I've done here is to use
implementation dependencies everywhere I could, and only switch to an
API dependency where required for compilation to succeed.  For
example, if class A in library Alpha is a subclass of class B in class
Beta, then code that tries to create a new A will fail unless Beta was
also brought along as an API dependency of Alpha.

In all cases where an API dependency was required, I have included an
explanation identifying at least one specific reason why, such as
"public class A extends class B".  Such explanations are optional, but
seem like a good idea and are even directly supported by Gradle's
dependency-declaration API.

It is definitely possible that some current implementation
dependencies should really be API dependencies.  This would most
likely happen if third-party code uses WALA APIs in ways that WALA's
own tests do not.  Returning to the A/Alpha/B/Beta example above, if
no WALA test actually uses class A, then we would we unaware of the
need for Alpha to propagate an API dependency on Beta.  A third-party
application or library code can work around such a misclassification
by explicitly adding its own dependency on Beta, so this is not a
show-stopping error.  It just makes WALA a bit less convenient to use
by requiring more explicit listing of dependencies that previously
would have been picked up implicitly.  We can certainly reclassify
implementation dependencies as API dependencies wherever appropriate
in the future.