Update linter plugin for Gradle scripts
Also update our configuration of this plugin to reflect the fact that
it does not currently understand implementation/api dependencies as
used with the java-library plugin.  We can turn these checks back on
if <https://github.com/nebula-plugins/gradle-lint-plugin/issues/203>
ever gets fixed, but unfortunately nobody is even working on that
right now.