Configure IntelliJ IDEA to build nearly all tests
This IntelliJ IDEA run configuration uses the Gradle "build" task to run
nearly all tests, excluding those tagged as slow.  It also runs all
Gradle-configured linters and builds Javadoc documentation.  This
configuration is intended to be a fairly thorough check that everything
is in order, while still not taking an unreasonably long time to finish.