Convert to Bash and use real Bash arrays
shellcheck was warning about sketchy word expansion.  At heart, $files
is really a set of changed file names.  We test this set for emptiness,
we emit this set's elements as a comma-delimited string, and we use this
set's elements as multiple subcommand arguments.  All of these are
easier to do robustly using a Bash array rather than keeping $files as a
flat string.