Allow raw use of generics in a few stubborn files
The raw use of generics in these files are vexingly hard to fix; I've
tried several times over the years and they just don't want to play
nice.  There's a very good chance that the code won't type-check because
it is, in fact, incorrect.  But as far as I'm aware nobody is triggering
failures and nobody other than me is trying to make these raw uses go
away.  So I'm going to categorize these in their own analysis scope and
suppress the raw-use-of-generics warning for that scope.  If I can't fix
them, then I don't want their warnings to be distracting me from other
things I can fix.