Suppress a legitimate shellcheck warning
My approach to fixing this was to port this script to Bash, but that's
not practical for other reasons discussed at length in
<https://github.com/wala/WALA/pull/599>.  I'm done spinning my wheels
here, so I'm just suppressing the warning.  It is probably a
legitimate warning, but would only affect WALA developers who (1) use
this script, and (2) keep their WALA tree in a path that included
spaces.  I am not such a developer, so I'm out.