Disable shellcheck linting on Windows
shellcheck is available for Windows, and easy to install.  However, it
expects the scripts that it checks to use Unix line endings (LF),
not Windows line endings (CRLF).  Git helpfully converts line endings to
Windows when checking out on Windows, then back to Unix when committing,
and in general that's a very good idea.  Disabling that conversion only
for shell scripts seems like a can of worms I don't want to open.