Reduce the scope of Maven builds (#618)
Maven builds now only do close to the minimum necessary to run the `com.ibm.wala.ide.*` Eclipse plug-in tests.  Unfortunately we can't easily port these test runs to Gradle as they rely on [Tycho](https://www.eclipse.org/tycho/), which is Maven-only.

As part of this change we do the following:
1. Remove most sub-modules from the top-level `pom.xml` so Maven no longer builds them.
2. Change the `com.ibm.wala.ide.*` projects to rely on the other WALA projects as Maven artifacts.  This means that `./gradlew install` must be run first for a Maven build to succeed or to see updated versions of other WALA code.  (Maybe we should run this as part of the Maven build?)
3. Related to 2, update the relevant Eclipse plug-in metadata in `com.ibm.wala.ide.*`.  In particular, we need exactly one plug-in to manage each core WALA Maven artifact jar and re-export its relevant packages.  Probably more packages are exported than needed right now; we can trim it down later if needed (I doubt it matters).
4. Update relevant Travis CI scripts in light of 2.

With these changes, a Maven build on Travis CI with a warm cache runs in under 5 minutes.

Note that we still pull in test data for `com.ibm.wala.ide.jdt.test` in a manner outside the WALA Maven artifacts, due to the particular structure required.  (`wala.cast.java.test.data` must be zipped in the structure of an Eclipse project and that zip needs to be copied over.)  We can further clean this up in the future.

This change is an intermediate step toward #462 and retiring Maven, and reduces the scope of Maven usage, so it is worth landing on its own.  A follow-up PR will remove various Maven, Ant, and Eclipse build configuration files that are no longer needed.