Treat com.ibm.wala.cast.java.test code as test fixture code
Previously this code was in the "main" sourceSet so that test code in
other subprojects could depend on it, and so that it would be published
among WALA's Maven Central artifacts.  See commit
4ed0970896ffab8ebff8186095e302c8960d3b43 and GitHub issue
<https://github.com/wala/WALA/pull/612> for that change and its intent.

Treating this as test fixture code better reflects its real role:
utility code used by tests.  This should fix
<https://github.com/wala/WALA/issues/639> while still satisfying the
needs of <<https://github.com/wala/WALA/pull/612>.