Use Java modules for anything except for Java 1.8
Under Java 1.8, we use the "sun.boot.class.path" system property to find
the boot class path.  But that went away in Java 9, and there's no
reason to think it's coming back any time soon.  Better to default to
the "$JAVA_HOME/jmods" approach, since that seems to be the established
standard going forward.