Log stdout and stderr from ECJ run
The "run" task here really serves as a test.  We don't validate the
output; the test merely consists of completing the run successfully.
So now we log this "run" task's stdout and stderr to a file, and set
that file as the tasks' output.  Without this, Gradle thinks that the
"run" task must be rerun every time we run the test suite.  In
general, Gradle caching and incremental rebuilds work best if it knows
where each task's output is stored.