Fetch upload details from project properties
Previously we were effectively looking into a Maven repository object
for the URL, username, and password to be used when uploading.  But
those details are supposed to be settable in, e.g.,
"~/.gradle/gradle.properties".  That means the properties we need will
be project properties, not properties on the Maven repository itself.