Move "hello_hash.jar" out of "build" and commit to repo
We use OCaml-Java to build "com.ibm.wala.core/ocaml/hello_hash.jar".
But downloading OCaml-Java is slow and flaky.  So we've decided to
commit the generated "hello_hash.jar" to the source repository.  If
that file exists, then we skip regenerating it, which transitively
means we skip downloading OCaml-Java.

If "hello_hash.jar" is missing, then we still download OCaml-Java and
rebuild "hello_hash.jar" as before.  Additionally, running "./gradlew
:com.ibm.wala.core:generateHelloHashJar" will rebuild this file if
needed, such as if it is missing or is older than any of its
transitive dependencies.

I initially thought that Task.outputs.upToDateWhen or Task.onlyIf
would be a suitable way to implement this, but neither did what I
wanted:
<https://discuss.gradle.org/t/short-circuit-dependency-chain-if-task-outputs-exist/35101>.

Fixes #710.