Improved JDK selection on macOS
Using `jdk:` in the Travis CI configuration file seems to have stopped
working.  What does work is to use `osx_image:` to pick a base image
that includes a suitable JDK as part of Xcode.

Also improve detection of Java 1.8 versus newer versions.  Previously
we checked $TRAVIS_JDK_VERSION, with a note that this is unset on
macOS with JDK 1.8.  Now it seems to be unset everywhere, always.  So
instead we check the Java version by running "javac -version" and
pattern-matching on that command's output.