Rely on "kawa.jar" using implicit dependencies
Putting "kawa.jar" on the classpath for dependent compilation tasks is
enough to establish the dependency, provided we do so in a way that lets
Gradle keep track of which other task is responsible for creating
"kawa.jar" in the first place.  That is, by connecting one task's
outputs to other tasks' inputs, we establish the dependency implicitly
instead of using explicit "dependsOn" directives.