Operate on Javadoc tasks now, not just eventually
Using `configureEach` here meant that we were installing callbacks to
be used eventually, when the subproject Javadoc tasks were
created and configured.  But that's not what we need here.  Rather, we
need to grab information from each of those Javadoc tasks right now, in
order to configure the `aggregatedJavadocs` task.