Ignore failed deletions during cache cleanup
This script is failing repeatedly on macOS when run during GitHub
Actions CI jobs.  The first "find ... -delete" command reports "No
such file or directory".  This is probably some transient
inconsistency in our view of the file system that holds the per-user
Gradle cache.  It should be safe to ignore those errors: if a file we
wanted to delete is already gone, then mission accomplished!