Suppress ECJ errors on Java test subjects
Previously we were suppressing these by adding "-err:none" to the ECJ
command line.  But that's not actually a valid ECJ flag.  It did
suppress the error diagnostics, but also produced a warning about the
invalid flag.  Now we suppress the specific error diagnostics that the
Java test subjects trigger, and we do it using valid ECJ flags.