Copy dependencies' jar files to where Eclipse expects them
Having these available to Eclipse resolves four "Plug-in problem"
reports concerning "build entry is missing" in
"com.ibm.wala.ide/build.properties".  There are still plenty of other
PDE-related problems reported, but perhaps this is a step in the right
direction.

Partially addresses #742, but certainly does not fix it completely.