Don't use gradle's temporaryDir for retained files
temporaryDir is supposed to be for short-lived files that need not
outlive a single task.  In theory, Gradle could delete these temporary
directories at any time, even though it currently does not.

A good example of something that would be sensible to store in
temporaryDir would be a file that contains additional command-line flags
to be given to an external command.  That file need not be retained
after the task that runs the command finishes.

By contrast, files that constitute a task's outputs should not go in
temporaryDir.  There's no guarantee that temporaryDir files will be
retained after each task finishes.