Introduce dependency locking (#900)
This was the simplest way I could find to introduce [dependency locking](https://docs.gradle.org/current/userguide/dependency_locking.html), which should prevent issues like #898 / #899 in the future.  It adds a bunch of empty lock files for configurations without dependencies, but since they are auto-generated I don't think this is a big deal; it will be more fragile to try to only generate non-empty lock files.

For now, we use the `LENIENT` lock mode to work around failure of certain checks in the default mode; see the comments for details.  Even in `LENIENT` mode, dynamic dependencies should be fixed at the versions in the lock files.