`compileTestJavaUsingEcj` depends on `downloadJLex`
`compileTestJavaUsingEcj` uses an output file created by `downloadJLex`,
but the former did not previously declare the latter as a dependency.
By adding that dependency now, we resolve a Gradle warning, enable
certain Gradle execution optimizations, and remove a potential source of
task-order-dependent non-determinism in builds.

Omitting dependencies in this manner is deprecated in Gradle 7 and will
be removed in Gradle 8.