Don't compile JSDT projects with Gradle (#944)
The JSDT Gradle projects pull artifacts from the Alfresco Maven repository, which now does not work due to an expired certificate.  Rather than trying to fix this in Gradle, this PR disables Gradle on the modules.  The modules are still built and tested via the "Maven Eclipse tests" CI job, which pulls the relevant artifacts from a p2 repository (via Tycho) so it is unaffected.