Give `autoLintGradle` task an empty output file
`autoLintGradle` creates no output files, which causes Gradle to treat
it as always out-of-date.  By creating a simple, empty stamp file to
record that this task has run and succeeded, we allow Gradle to avoid
rerunning this task unnecessarily.  This task will still be rerun when
needed, though, such as when any of the `**/build.gradle` files change.