Some updates to Gradle lock files (#952)
These were generated by running `./gradlew allDeps --write-locks`.  The changes include:

1. A new lock file for `com.ibm.wala.core.java11`
2. Further updates for JUnit and Hamcrest dependencies that were missed in #939 
3. Various updates to Eclipse-related dependencies that I don't fully understand, but might be due to a changed version of the `com.diffplug.eclipse.mavencentral`

It's a bit puzzling to me that the CI jobs were all passing despite the issues in 2 and 3.  My best guess is that the build will only fail for _dynamic_ dependencies of configurations that get resolved, where the resolved dependence doesn't match the lock file.  So for 2 and 3, maybe the relevant configurations are never resolved on CI, or maybe these are changes in non-dynamic dependencies.

FWIW, this change doesn't seem to affect dependencies showing up in https://github.com/wala/WALA/issues/937#issuecomment-913363643; so that must be a separate issue.