Revive IFDSExplorerExample (#975)
Fixes #921 

Previously we did not test that the SWT-related code in `com.ibm.wala.ide.tests` actually ran correctly, and we ended up with some classpath issues not detected by just building the code (e.g., a dependence that required Java 11).  This PR fixes these issues, and adds a unit test so we don't regress in the future.

Also, we add a Gradle task to run `IFDSExplorerExample` to make it easier for others to try out in the future.