Double-quote a variable expansion
This variable expansion is actually safe even without quoting, since it
can only ever be a number from 1 through 10.  But ShellCheck apparently
isn't smart enough to know that.  In general, we want to keep this kind
of warning enabled, so the easiest thing to do is add the (unnecessary)
quotes here to satisfy ShellCheck.