Cancel outdated jobs using workflow concurrency (#989)
This is an improved version of #976 using [the `concurrency` feature](https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#concurrency) of GH Actions workflows.  This ensures our CI workflow only runs for the latest commit on any branch / PR, and also that only one workflow for each branch is running at any given time.  The latter is useful for the WALA master branch in particular, since before it was possible for workflows to upload Javadoc and snapshot builds out of order if one ran particularly slowly.