Annotate some methods as JUnit 4 tests
These all seem to be test methods, but were lacking `@Test` annotations.
For several, a `@Test` annotation had clearly been commented out,
sometimes with an explanatory comment.  The `@Ignore` annotation is a
better way to let a test method be marked as a test method, but not run.
That annotation even accepts an optional string argument to explain why
the annotated test method is being skipped.