Run all CI jobs on JDK 11 (#1026)
It seems that with the current Gradle config, all Gradle tasks getting run on JDK 8 on Linux also pass on JDK 11.  So, this PR enables them for both versions.  Note that there is still some work to do to get Dalvik tests running on JDK 11, but they are disabled in the Gradle config for `com.ibm.wala.dalvik` on JDK 11 so this change still works.

We also update the JDK distribution to "temurin", the new name for AdoptOpenJDK:

https://github.com/actions/setup-java#supported-distributions