No-op ordering changes to IntelliJ IDEA inspections configuration
I have not actually changed the configuration of this "No Back Sliding"
inspections profile.  But as of IntelliJ IDEA 2021.2.3, the order in
which entries are written out has changed.  OK, we should let IntelliJ
IDEA do what it wants here, so that any future *intentional* changes to
this profile will be easier to understand.