Suppress a warning about side-effects in an assertion
Surprisingly, `BitVectorIntSet::isEmpty` can actually modify the
receiving instance:  `BitVectorIntSet::isEmpty` calls
`BitVectorIntSet::size`, which might assign a refreshed value to
`BitVectorIntSet.populationCount`.  IntelliJ IDEA infers contracts for
both `BitVectorIntSet::size` and `BitVectorIntSet::isEmpty` that reflect
this possible modification of `this`, which in turn leads to a warning
about side effects in an `assert` statement.  In this particular case,
though, that possible side effect is harmless.