Build against Eclipse 4.21.0 on ARM-based Macs (#1047)
This change allows for Eclipse-based code that uses SWT to build on ARM-based Macs. A more recent Eclipse version is needed since an SWT artifact for ARM Mac was only introduced recently. We do not bump the Eclipse version globally since recent Eclipse versions require JDK 11, and we want to preserve the ability to run this code on JDK 8.