Stop using external gpg command (#1059)
It seems the external gpg command support is poor and doesn't really work out of the box with `gpg2`.  I have re-configured the signing config on my local machine to be compatible and `gpg2` and the built-in Gradle signing functionality.  So this is no longer necessary.  I will also update the documentation at https://github.com/wala/WALA/wiki/WALA-Releases-and-Snapshots#initial-setup appropriately.

I've tested locally and signing seems to work, though the true test will only happen next time we want to cut a release (which should probably be soon anyway).