Download JLex into build subdir, not sources
The `JLex.java` file that we download is compiled as part of the
`com.ibm.wala.test.data` test sources, but we still shouldn't treat it
as "real" hand-authored, git-tracked source code.  This file is built by
running a task, just like any other build artifact.

Happily, treating `JLex.java` as a built actually makes the relevant
build script simpler.  There's some subtlety to the way that this
download gets hooked up as source to compile, but now that we know how
to do that, this change lets us shrink the build logic by about a dozen
lines.