Fix SourceMapTest so it can run on CI (#1064)
In order for the `SourceMapTest` tests to pass, they must be able to see files inside the wala.core `testSubjects` source directory.  This PR adjusts the `testRuntimeOnly` path to include this directory, so the tests now pass.  This change enables removing the `analyzingJar` configuration that was added to skip the tests.