Refine gathering of classpath for whole-project Javadoc
Our previous approach triggered deprecation warnings.  This new approach
works fine, and should be compatible with some future Gradle 8.x
release.  The important change seems to be to treat subprojects' Javadoc
tasks' class paths as Gradle "artifacts", which are allowed to be
produced and consumed in different (sub)projects.

Fixes #1079.