Cache all predecessors in CHACallGraph (#1125)
Previously, computing the predecessors of each node in the `CHACallGraph` would require traversing all call graph nodes, a very expensive operation.  This PR changes the implementation to store the predecessors in a data structure from the beginning, making predecessor lookup fast.  This consumes more memory, but the benefits in running time are worth it.