Make demand-driven points-to analysis robust to missing call sites (#1127)
This can arise in rare cases where the up-front call graph is a `CHACallGraph`, which may consider call sites not present after constructing IR (which does some dead-code elimination).  It's hard to create a test for this scenario, but the code changes are safe, in that they are adding nullness / array length checks that were not present before.