Revamp classpath search to use normalized paths
On Windows with Java 11, the `java.class.path` property contains
backslash-escaped backslashes, as in "`C:\\foo\\bar`".  These doubled
backslashes interfere with the logic for finding classpath entries that
contain the requested element.  Using normalized paths for all
comparisons seems to fix the problem, and will likely be more robust in
general.